requirements.txt
streamlit
Pillow
numpy
requests
